[level 0] Title: 두 수의 나눗셈, Time: 0.04 ms, Memory: 87.9 MB -BaekjoonHub

+++ 프로그래머스/0/120806. 두 수의 나눗셈/두 수의 나눗셈.java
@@ -1,7 +1,7 @@
 class Solution {
     public double solution(int num1, int num2) {
         double answer = 0;
-        answer = (int)(((double)num1 / num2) * 1000);
-        return answer;
+        answer = (double)num1 / (double)num2 * 1000;
+        return (int)answer;
     }
 }
